moves all tests to tests/old-structure
