+++ __tests__/old-structure/BeagleAnalytics.spec.ts
