+++ __tests__/old-structure/BeagleHttpClient.spec.ts
