+++ __tests__/old-structure/BeagleUIView.spec.ts
