+++ __tests__/old-structure/GlobalContextAPI.spec.ts
