+++ __tests__/old-structure/actions/confirm.spec.ts
