+++ __tests__/old-structure/actions/navigation.spec.ts
