+++ __tests__/old-structure/actions/sendRequest.spec.ts
