+++ __tests__/old-structure/globals.d.ts
