+++ __tests__/old-structure/middlewares/beagleShouldPrefetch.spec.ts
