+++ __tests__/old-structure/middlewares/beagleStyleClassMiddleware.spec.ts
