+++ __tests__/old-structure/mocks-global-context.ts
