+++ __tests__/old-structure/styles-mocks.ts
