+++ __tests__/old-structure/utils/string.spec.ts
